Building és upgrade effektek implementálva

+++ backend/StrategyGame.Model/Country.cs
@@ -27,8 +27,8 @@ namespace StrategyGame.Model
         public List<Unit> Units { get; set; }
         public int Population { get; set; }
         public int ArmyCapacity { get; set; }
-        public int AttackModifier { get; set; } = 1;
-        public int DefenseModifier { get; set; } = 1;
+        public double AttackModifier { get; set; } = 1;
+        public double DefenseModifier { get; set; } = 1;
 
     }
 }
